+++ src/model/layers/initialization.cpp
@@ -9,7 +9,7 @@ void GolorotInitialization::initialize(
     std::srand(time(NULL));
     double limit_ih = std::sqrt(6.0/dim_sum);
 
-    for(int i=0; i<tensor.size(); i++){
+    for(size_t i=0; i<tensor.size(); i++){
         tensor[i] = ((double) std::rand() / RAND_MAX) * 2 * limit_ih - limit_ih;
     }
 }
